Added buttons and labels and business logic description
I didn’t have time to implement the algorithm I outlined, but it
shouldn’t be hard to do. As suggested in the comment, if anyone has a
better/different idea of how to do the business logic, feel free to
implement it. You won’t hurt my feelings. Email me with any questions.